backend/functions/publish_event/requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.18.0
